// src/generated/model/productMealInfoModel.ts
/**
 * Datafood Server API
 * API contract of Datafood server application BRASIDAS
 *
 * The version of the OpenAPI document: 0.1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface ProductMealInfoModel { 
    idProduct?: string;
    labelProduct?: string;
    nutriscore?: string;
    imageProduct?: string;
    quantity?: number;
    packagingQuantity?: number;
    mealId?: string;
    mealType?: string;
    idLProductMeal?: string;
    date?: string;
}